requirements.txt
asgiref==3.2.10
boto3==1.17.43
botocore==1.20.43
certifi==2020.6.20
chardet==3.0.4
Django==3.1
django-cors-headers==3.5.0
django-filter==2.3.0
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
drf-spectacular==0.17.2
idna==2.10
install==1.3.3
jmespath==0.10.0
Markdown==3.2.2
# mysqlclient==2.0.1
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
s3transfer==0.3.6
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
